delete needless code and combine similar functions
