+++ src/js/markup/markup.js
@@ -6,9 +6,9 @@ import {
   getSearchArticle,
   getDataByCategory,
 } from '../api/news';
-import { createNewsCard, newsCardTextFormat } from '../newsCard/newsCard';
+import { createNewsCard } from '../markup/card';
+import { Notify } from 'notiflix/build/notiflix-notify-aio';
 
-let arrayNewsCard = [];
 const error =
   'https://img.freepik.com/free-vector/404-error-with-cute-animal-concept-illustration_114360-1931.jpg';
 let page = 2;
@@ -19,91 +19,93 @@ refs.filterCategories.addEventListener('click', renderNewsCategory);
 
 function saveValuesFromCategoryNews(articles) {
   console.log(articles);
-  articles.map(article => {
-    arrayNewsCard.push({
-      title: article.title,
-      media: `${
-        article.multimedia === null ? error : `${article.multimedia[3].url}`
-      }`,
-      url: article.url,
-      published_date: article.published_date,
-      section: article.section,
-      abstract: article.abstract,
+  return articles.map(article => {
+    return {
       id: article.id,
+      media: `${
+        article.multimedia?.[3]?.url ? article.multimedia?.[3]?.url : error
+      }`,
+      title: article.title,
+      abstract: newsCardTextFormat(article.abstract),
+      section: article.section,
+      published_date: article.published_date,
+      url: article.url,
       uri: article.uri,
-    });
+    };
   });
 }
 
 function saveValuesFromSearchNews(articles) {
-  articles.map(article => {
-    arrayNewsCard.push({
-      title: article.headline.main,
-
-      media: `${
-        article.multimedia[0] === undefined
-          ? error
-          : `https://static01.nyt.com/${article.multimedia[0].url}`
-      }`,
-
-      url: article.url,
-      published_date: article.pub_date,
-      section: article.section_name,
-      abstract: article.abstract,
+  return articles.map(article => {
+    return {
       id: article._id,
+      media: `${
+        article.multimedia?.[0]?.url
+          ? `https://static01.nyt.com/${article.multimedia[0].url}`
+          : error
+      }`,
+      title: article.headline.main,
+      section: article.section_name,
+      abstract: newsCardTextFormat(article.abstract),
+      published_date: article.pub_date,
+      url: article.url,
       uri: article.uri,
-    });
+    };
   });
 }
+
 function saveValuesFromPopularNews(articles) {
-  articles.map(article => {
-    arrayNewsCard.push({
-      title: article.title,
-      media: `${
-        article.media[0] === undefined
-          ? error
-          : article.media[0]['media-metadata'][2].url
-      }`,
-      url: article.url,
-      published_date: article.published_date,
-      section: article.section,
-      abstract: article.abstract,
+  return articles.map(article => {
+    return {
       id: article.id,
+      media: `${
+        article.media?.[0]?.['media-metadata']?.[2]?.url
+          ? article.media?.[0]?.['media-metadata']?.[2]?.url
+          : error
+      }`,
+      title: article.title,
+      section: article.section,
+      abstract: newsCardTextFormat(article.abstract),
+      published_date: article.published_date,
+      url: article.url,
       uri: article.uri,
-    });
+    };
   });
+}
+
+function newsCardTextFormat(element) {
+  let textFormat = element;
+  if (textFormat.length > 80) {
+    textFormat = element.slice(0, 80) + '...';
+  }
+  return textFormat;
 }
 
 function renderPopularNews(articles) {
-  refs.newsList.innerHTML = '';
-  arrayNewsCard = [];
+  const cardArray = saveValuesFromPopularNews(articles);
 
-  saveValuesFromPopularNews(articles);
-  renderNewsList(arrayNewsCard);
+  resetNewsList();
+  renderNewsList(cardArray);
 }
 
 function renderSearchNews(e) {
   e.preventDefault();
-
-  refs.newsList.innerHTML = '';
-  arrayNewsCard = [];
+  resetNewsList();
 
   const date = refs.celendarDate.dataset.time.replaceAll('-', '');
-  console.log(date);
-
   const inputSearchValue = refs.form.elements.inputSearch.value;
+
   getSearchArticle(inputSearchValue, page, date)
     .then(articles => {
-      // console.log(articles);
-      saveValuesFromSearchNews(articles);
-      renderNewsList(arrayNewsCard);
+      const cardArray = saveValuesFromSearchNews(articles);
+      renderNewsList(cardArray);
     })
     .catch()
     .finally(data => {
       hideLoader();
-      // reset()
     });
 }
+
 function renderNewsCategory(e) {
   console.log(e.target);
   console.log(refs.filterOthers);
@@ -112,15 +114,17 @@ function renderNewsCategory(e) {
   }
   const categoryName = e.target.dataset.category_name;
 
-  console.log(categoryName);
-
   getDataByCategory(categoryName)
     .then(articles => {
       refs.newsList.innerHTML = '';
-      arrayNewsCard = [];
+      const cardArray = saveValuesFromCategoryNews(articles);
 
-      saveValuesFromCategoryNews(articles);
-      renderNewsList(arrayNewsCard);
+      if (cardArray.length < 1) {
+        Notify.failure('Error: No news found');
+        return;
+      }
+
+      renderNewsList(cardArray);
     })
     .catch()
     .finally(hideLoader());
@@ -133,22 +137,28 @@ function renderNewsList(arrayNewsCard) {
       createMarkupWidgetWeather(orderedNumber) + previousValue;
     }
 
-    return createNewsCard(article, orderedNumber) + previousValue;
+    return createNewsCard(article) + previousValue;
   }, '');
+
   updateNewsList(markup);
   orderedNumber = 0;
 }
 
 function updateNewsList(markup) {
-  // refs.newsList.insertAdjacentHTML('beforeend', markup);
   refs.newsList.innerHTML = markup;
 }
+
+function resetNewsList() {
+  refs.newsList.innerHTML = '';
+}
+
 function createMarkupWidgetWeather() {
   return `<li class =" news__item location_weather"  ><div class=" news__weather"><p class = "text_weather">Weather<p></div></li>`;
 }
+
 export {
   renderNewsList,
-  updateNewList,
+  updateNewsList,
   createMarkupWidgetWeather,
   orderedNumber,
   renderPopularNews,
